services/api/requirements.txt
fastapi==0.115.6
mangum==0.19.0
boto3==1.35.95
python-jose[cryptography]==3.3.0
pydantic==2.10.3
pydantic-settings==2.7.0
uvicorn==0.34.0
